# apps/api/src/booking/dto/create-booking.dto.ts
import { IsString, IsDateString } from 'class-validator';

export class CreateBookingDto {
  @IsString()
  userId: string;

  @IsString()
  dealershipId: string;

  @IsDateString()
  scheduledTime: string;
}