# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered."},"409":{"description":"Email already exists."}},"summary":"Register a new user","tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in."},"401":{"description":"Invalid credentials."}},"summary":"Login user","tags":["Auth"]}},"/vehicles":{"get":{"operationId":"VehicleController_findAll","parameters":[],"responses":{"200":{"description":"List of all vehicles."}},"summary":"Get all vehicles","tags":["Vehicles"]}},"/vehicles/{id}":{"get":{"operationId":"VehicleController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Vehicle details."},"404":{"description":"Vehicle not found."}},"summary":"Get vehicle by ID","tags":["Vehicles"]}},"/bookings":{"post":{"operationId":"BookingController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":""}},"tags":["Booking"]}},"/customizations":{"post":{"operationId":"CustomizationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomizationDto"}}}},"responses":{"201":{"description":"Customization saved."}},"security":[{"bearer":[]}],"summary":"Save a customization","tags":["Customizations"]},"get":{"operationId":"CustomizationController_findAll","parameters":[],"responses":{"200":{"description":"List of customizations."}},"security":[{"bearer":[]}],"summary":"Get my customizations","tags":["Customizations"]}},"/customizations/{id}":{"get":{"operationId":"CustomizationController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Customization details."},"404":{"description":"Customization not found."}},"security":[{"bearer":[]}],"summary":"Get customization by ID","tags":["Customizations"]}},"/orders":{"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"Order created."}},"security":[{"bearer":[]}],"summary":"Create an order from customization","tags":["Orders"]},"get":{"operationId":"OrderController_findAll","parameters":[],"responses":{"200":{"description":"List of orders."}},"security":[{"bearer":[]}],"summary":"Get my orders","tags":["Orders"]}}},"info":{"title":"Velocity API","description":"The Velocity Automotive Commerce API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@velocity.com"},"password":{"type":"string","example":"password123","minLength":6},"name":{"type":"string","example":"John Doe"}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@velocity.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"CreateBookingDto":{"type":"object","properties":{}},"CreateCustomizationDto":{"type":"object","properties":{"vehicleId":{"type":"string","example":"vehicle-uuid"},"config":{"type":"string","example":"{\"COLOR\": \"option-uuid\", \"WHEELS\": \"option-uuid\"}"},"totalPrice":{"type":"number","example":50000},"preview":{"type":"string"}},"required":["vehicleId","config","totalPrice"]},"CreateOrderDto":{"type":"object","properties":{"customizationId":{"type":"string","example":"customization-uuid"}},"required":["customizationId"]}}}}